+++ src/p2p/p2p_server.h
@@ -4,7 +4,7 @@
 #include <memory>
 #include <string>
 
-// nlohmann/json
+// Boost
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>
 #include <boost/system/error_code.hpp>