+++ src/p2p/p2p_session.h
@@ -6,7 +6,7 @@
 #include <memory>
 #include <string>
 
-// nlohmann/json
+// Boost
 #include <boost/asio/bind_executor.hpp>
 #include <boost/asio/io_context.hpp>
 #include <boost/asio/ip/tcp.hpp>